test(esl-event-listener): refactor for listener methods
